requirements.txt
python-telegram-bot>=21.9
python-dotenv
requests>=2.31.0
pycryptomusapi>=1.0.0
mysql-connector-python>=8.2.0
